+++ database/migrations/2021_03_18_084248_user_add_column_wallet.php
@@ -14,7 +14,7 @@ class UserAddColumnWallet extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {  
-           $table->float('lock_wallet_balance')->after('wallet_balance')->default(0);          
+           $table->double('lock_wallet_balance')->after('wallet_balance')->default(0);          
            $table->string('user_timezone')->after('lock_wallet_balance')->nullable();       
         });
     }
