+++ database/migrations/2021_05_17_061742_chat_history_add_amountfield.php
@@ -14,7 +14,7 @@ class ChatHistoryAddAmountfield extends Migration
     public function up()
     {
         Schema::table('chat_histories', function (Blueprint $table) {            
-            $table->float('transaction_amount')->nullable(); 
+            $table->double('transaction_amount')->nullable(); 
             
         });
     }
